Add totales en AdmDocumentosOp,AdmImpuestosDocumentosOp

+++ Convertidor/Dtos/Adm/AdmImpuestosDocumentosOpResponseDto.cs
@@ -9,10 +9,10 @@
         public int TipoRetencionId { get; set; }
         public string DescripcionTipoRetencion { get; set; } = string.Empty;
         public string PeriodoImpositivo { get; set; } = string.Empty;
-        public int BaseImponible { get; set; }
-        public int MontoImpuesto { get; set; }
-        public int MontoImpuestoExento { get; set; }
-        public int MontoRetenido { get; set; }
+        public decimal BaseImponible { get; set; }
+        public decimal MontoImpuesto { get; set; }
+        public decimal MontoImpuestoExento { get; set; }
+        public decimal MontoRetenido { get; set; }
    
     }
 }
